Doc: Move and edit examples page
Changed the example page type to a group to have a standardized
table of examples. Also, a filename 'examples.html' is too
collision-prone when Qt docs are output to a single directory.

Also, moved the examples.qdoc file into /doc subdirectory to
avoid it from being copied to QT_INSTALL_EXAMPLES when packaging
the examples.

Change-Id: I87251e0fde127bdef3024907b6472b50039f0357
Reviewed-by: Kurt Pattyn <pattyn.kurt@gmail.com>