Add initial support for a declarative WebSocketServer.
It comes with basic support for setting up a non-secure server
listening on an arbitrary host and port. The name can be set,
and whether the server should be listening or accept connections.
The error string is also available via a property.

When a client connects, a clientConnected(QQmlWebSocket*) signal is
emitted.

To showcase the usage and functionality, a new qmlwebsocketserver
example is added which behaves similar to qmlwebsocketclient. It
simply sends messages from a WebSocket to a local WebSocketServer
which in turn sends a simple response.

Change-Id: Ib6318155ed08cbeea47d820b709764d2e36dcc45
Reviewed-by: Kurt Pattyn <pattyn.kurt@gmail.com>