Fixed QWebSocketServer::SslMode collision
Value of NonSecureMode enumerator depends on QT_NO_SSL. If Qt is built with
QT_NO_SSL defined, NonSecureMode is equal to 0, otherwise it is equal to 1.
It may lead to problem if user built app with libs where NonSecureMode = 0,
but run it with libs where NonSecureMode = 1.

[ChangeLog][QWebSocketServer] SslMode::NonSecure enumerator now always has
the same value (1) regardless of SSL enabled or disabled.

Task-number: QTBUG-56129
Change-Id: If17137eff4b19015697c827a985ef89dd95a5611
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>