QQmlWebSocket::setUrl(): let open() decided what to actually do
After emitting urlChanged(), setUrl() did a manual open(url) if it had
a web-socket to do it on, omitting various checks that open() imposes
on actually calling open(url); as a result, it left the web socket in
an inconsistent state if it wasn't yet active.  This meant that
setting active could trigger a second open() while the web-socket was
already open(), which upset it.  Replaced the hand-rolled call to
open(url) with a call to open() so as to be consistent and avoid such
problems.

Task-number: QTBUG-58278
Change-Id: Iee84f2b0d973e7fd288a14018039a665658c0040
Reviewed-by: Jesus Fernandez <Jesus.Fernandez@qt.io>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>