Divide QWebSocketFrame::readFrame into subroutines
To reduce the dependency of each state and remove local variables
(bytesRead, hasMask, and payloadLength) from QWebSocketFrame::readFrame,
this change breaks down the states of QWebSocketFrame::readFrame into
subroutines. In this change, it uses a readFramePayloadLength function
to handle PS_READ_PAYLOAD_LENGTH and PS_READ_BIG_PAYLOAD_LENGTH cases.
In addition, a missing error-handling is added in the read-header case
if the bytesAvailable is less than 2 bytes.

The variable hasMask is replaced by transiently using m_mask to
indicate whether to read a mask. After reading mask in readFrameMask,
the transient value will be over-written and used in readFramePayload.

To replace local variable payloadLength with QWebSocketFrame::m_length,
m_length is expanded to hold the payload length.

Change-Id: I1c2d197112cef5cea75215923fa28e6f2e6cbef8
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>