Investigate tst_handshakeTimeout test flakiness
Remove "localhost" in favor of "127.0.0.1" to avoid any name resolution.

And wait 5 seconds for all websockets to timeout together, while
printing how many have not timed-out in case of failure.
Credits to Mårten Nordheim for coding this.

With this patch, I can not reproduce the flakiness any longer.
Lets see if this is also the case in our CI.

Fixes: QTBUG-76572
Change-Id: Ibdbafbd7786c87138bdc84ebc690177211f23f5b
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>