wasm: make sendBinaryMessage work in threaded mode
This fixes two issues in sendBinaryMessage(), one
observed, and one theoretical:

1) WebSocket send() does not accept data views backed
by SharedArrayBuffer, which is the case for heap memory
views when threading is enabled.

2) We have no way of observing for how long send() will
retain the memory view passed to it. This means we
don’t know when the QByteArray can be safely freed.

Both can be solved by copying the payload data to a
separate ArrayBuffer, whose lifetime can be managed
by WebSocket.

Fixes: QTBUG-78078
Change-Id: I73209080db66f38b971f2c8a727b43402357b1a9
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>