Make sure child QObjects get moved-to-thread along with the owner
By parenting them. And to do that without crashing they need to be
allocated dynamically.

A couple missed "parenting"s caused the object to not be moved to the
other thread when the websocket was moved. This caused it to print
warnings when the timer was started and stopped since this cannot be
done across threads.

Fixes: QTBUG-83722
Change-Id: Iee60fe1c498f8f6d1e0c2cfcb2923bd5b9560acb
Reviewed-by: Jesus Fernandez <jsfdez@gmail.com>
(cherry picked from commit 393fcf8a64748f0c59f47d254667518dba9e2fc3)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>