Make sure child QObjects get moved-to-thread along with the owner
By parenting them. And to do that without crashing they need to be
allocated dynamically.

A couple missed "parenting"s caused the object to not be moved to the
other thread when the websocket was moved. This caused it to print
warnings when the timer was started and stopped since this cannot be
done across threads.

Fixes: QTBUG-83722
Pick-to: 5.15
Change-Id: Iee60fe1c498f8f6d1e0c2cfcb2923bd5b9560acb
Reviewed-by: Jesus Fernandez <jsfdez@gmail.com>