QSslServer - adopt the newly created socket
Like it's done in QTcpServer. In theory, it's possible that 'encrypted'
signal is never executed and then the socket is not added to the pending
connections (those are deleted in the QTcpServer::close).
Thus, the socket object  would be leaked. Since QSslServer inherits from
QTcpServer, this fix also ensures the documented behavior of the
QTcpServer::nextPendingConnection (which is said to return a child
of the server object).

Fixes: QTBUG-85180
Pick-to: 5.15
Change-Id: Ic0d923b7a7fe5e3bcea0b9815be1cbe946dda186
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>