Update commercial license headers
Updated header.COMM to the files in tqtc-qtwebsockets. Examples,
tests or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is
that some autotests use hard coded line numbers
and a change in the line count causes failures in tests.

Task-number: QTQAINFRA-4218
Change-Id: I91525398b24d696254f551f3db57107b5197e38f
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>