Doc: Update \externalpage links to RFCs
The old `tools.ietf.org/html/*` web pages containing RFCs now redirect
to `datatracker.ietf.org/doc/html/*`.

While the redirection work, our script to catch broken links on
dev-snapshots builds of the documentation doesn't handle redirection
correctly, reporting it as broken.

Both to appease the broken-links script and to avoid an unneccesary
redirection the RFC links in `external-resourced.qdoc` were modified to point
to the new domain.

Furthermore, the link referring to a specific page that pointed to a
section were updated to point directly to the section, so that they are
consistent with other similar links.

The `tools.ietf.org/id/*.txt` links, while still resolving without a
redirection, were updated to use an equivalent link on the
`datatracker.ietf.org` domain for consistency.

Task-number: QTBUG-96127
Pick-to: 6.2 6.2.0
Change-Id: I0b937c9dbfaec0cd4b2aad5bd55fc731d073ca16
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>