Doc: Repair broken link to a websocket paper
The `\externalpage` link to the "Talking to Yourself for Fun and Profit"
paper, presented at the Web 2.0 Security and Privacy 2011 conference,
pointed to the version hosted on the `w2spconf` domain, which is not
available anymore.

To repair the broken link, it was changed to point to an equivalent
version stored on the `ieee-security` site.

Task-number: QTBUG-96127
Change-Id: Id68c5947fa285062fa69a35fa9641e88118448ba
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit bfa705a7a20576405e77be28d185c84f96c1acd0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>