Doc: Correct a recently introduced typo in an \externalpage link
Commit bfa705a7a20576405e77be28d185c84f96c1acd0 recently modified the
link to the "Talking to Yourself for Fun and Profit" paper.

While the new location for the `\externalpage` link is correct, it was
erroneously introduced with a typo in it.

In particular, `fd` was introduced in front of the `org` domain such
that the link was incorrect.

The typo is now removed such as to correct the link.

Task-number: QTBUG-96127
Change-Id: Ie4d169a978767b77f6d59b2ef135b3a3a8a0131d
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 3e855cde66a60f57eb7b307c6e6760d58f208397)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>