Update commercial license headers
Updated header.COMM to the files in tqtc-qtwebsockets.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4962
Change-Id: I6ad9d843376efa1597eb3b76180ae3485129f8a0
Reviewed-by: Antti Kokko <antti.kokko@qt.io>