QtWebSockets: replace qSwap with member-swap where possible
qSwap() is a monster that looks for ADL overloads of swap() and also
detects the noexcept of the wrapped swap() function, so it should only
be used when the argument type is unknown. In the vast majority of
cases, the type is known to be efficiently std::swap()able or to have
a member-swap. Call either of these.

Task-number: QTBUG-97601
Pick-to: 6.3 6.2 5.15
Change-Id: If9c28df7d788856fb1653d2fae2a68ff42a12b23
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>