QWebSocketServer: update handshake timeout to keep up with QSslServer
A side-effect of this is that the timeout is now duplicated.
The TLS handshake first has the 'msec' timeout, and then the
websocket handshake has the same 'msec' timeout.

[ChangeLog][QWebSocketServer] Due to the introduction of
a standalone QSslServer, with its own timeout handling, the
setHandshakeTimeout() function now applies the same timeout
to both the TLS and WebSocket handshakes separately.

Pick-to: 6.4
Fixes: QTBUG-105851
Change-Id: I6c515e0dcdf83fa452979b06ab5f890c8b14b184
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
Reviewed-by: Ievgenii Meshcheriakov <ievgenii.meshcheriakov@qt.io>