Doc: Remove usages of the second argument of a "\page" command
The "\page" command for QDoc, used to construct a standalone
documentation page in the output documentation of a project, was able to
take a second argument, apart from the first argument representing the
name for the generated documentation page, that roughly represented the
type of the page that was to be generated.

This second argument was not actually used by QDoc, such that it had no
meaningful effect.

QDoc was recently modified to not support this second argument,
internally, and officially removed its use.

For technical legacy reason, QDoc will still support the usage of a
second argument for a "\page" command, albeit it will do nothing, as
before.

To avoid confusion for future readers who might not be aware of the
history of the "\page" argument and that will not find an explanation in
QDoc's documentation anymore, the usages of the second argument in
"\page" commands are now removed.

Change-Id: Ia41f57d5afd76b0c369f58b3f9c25854e55a33fc
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>