Address issues with test introduced with 401 support
The testing with Connection: Close wasn't necessarily testing it since it was
relying on not having a Content-Length, but this did not work out as intended.

And fix excluded code (#if 0), relies on an update in qtbase.
On macOS the ST backend was having trouble with how we handled shutdown of
socket and re-use.

Amends da30f70fea239f723f1d36b076bb3f5860f50ed9

Change-Id: I8a19387b911eec522690d597d863c5bc0786e8c8
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>