Expand documentation of examples
Add to the documentation to include what the examples do and how they
do it, to make it easier for users to decide if an example is worth
looking into, instead of how it was where they had to look at the code
in the examples to decide if they are relevant for their purposes.
Also added pictures to the examples.

Task-number: QTBUG-110894
Change-Id: I2751939e9f9716cd970e5fa938dabf042f369329
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit 0b665063815a429d7862e261dfa469ca50c55fa4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>