Doc: Use the \examplecategory macro to tag examples
The '\meta category' command was used for tagging examples with a
specific category, used in Qt Creators Welcome mode.

As we want to also generate lists of examples belonging to a category
elsewhere in the documentation, replace the command with a macro that
expands to the original \meta command and also adds the example to
a group using the \ingroup command. This way, the category names can
be used as arguments to the \generatelist or \annotatedlist commands.

Task-number: QTBUG-112731
Change-Id: Ia1b34175e10109eef055b2759705f0ca0521179a
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit a5f6f3f710fc9de3f8b9d4685e22c9c8912b11d9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>