Doc: Remove warninglimit .qdocconf variable
The most common limit for the maximum number of allowed documentation
warnings is zero. As all Qt module docs include() a common configuration
from qtbase, it's better to control the base warninglimit from a central
location in qtbase/doc/global.

This allows for a temporary increase of the limit across all modules as
needed - for example, when updating the QDoc binary that the CI
provisions to a version that introduces new types of documentation
warnings.

Task-number: QTBUG-113326
Change-Id: Ia31a66b11ea3ffe95747c901139bb960a757e4a6
Reviewed-by: Paul Wicking <paul.wicking@qt.io>