QWebSocketDataProcessor: Restore utf-8 decoding state
The data processing happens on frames, so the decoding needs to cope
with sequences on the boundaries of frames. That's what the state is for.

It was made stateless likely as an oversight during Qt 6 porting when
moving away from QTextCodec.

Amends aeeaa00fa083ac339d0770d633d22f203c504253

Pick-to: 6.9 6.8 6.5
Fixes: QTBUG-136216
Change-Id: I5492febc5bff9ed1f70f995ea7af73a1655c4811
Reviewed-by: Mate Barany <mate.barany@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>